Changes enrollements helper file name into enrollments_controller
