+++ app/controllers/discharges_controller.rb
@@ -1,9 +1,5 @@
-require 'enrollments_helper'
-
-class DischargesController < GeneralEnrollmentsController
-
+class DischargesController < EnrollmentsController
   def correct_params
     params["facility_discharged_date"] = params.delete "on"
   end
-
 end
