+++ app/controllers/placements_controller.rb
@@ -1,9 +1,5 @@
-require 'enrollments_helper'
-
-class PlacementsController < GeneralEnrollmentsController
-
+class PlacementsController < EnrollmentsController
   def correct_params
     params["date_orr_approved"] = params.delete "on"
   end
-
 end
