+++ app/controllers/referrals_controller.rb
@@ -1,9 +1,5 @@
-require 'enrollments_helper'
-
-class ReferralsController < GeneralEnrollmentsController
-
+class ReferralsController < EnrollmentsController
   def correct_params
     params["referral_date"] = params.delete "on"
   end
-
 end
